refactor: move projetForm in gsl_projet
